+++ starskyapp/src/app/ipc-bridge/ipc-bridge.spec.ts
@@ -53,9 +53,24 @@ describe("ipc bridge", () => {
         return Promise.resolve(true);
       });
 
-      jest.spyOn(appConfig, "set").mockImplementationOnce(() => {
-        return Promise.resolve();
-      });
+      jest
+        .spyOn(appConfig, "set")
+        .mockImplementationOnce(() => {
+          return Promise.resolve();
+        })
+        .mockImplementationOnce(() => {
+          return Promise.resolve();
+        });
+
+      jest
+        .spyOn(createMainWindow, "default")
+        .mockImplementationOnce(() =>
+          Promise.resolve({ once: jest.fn() } as any)
+        );
+
+      jest
+        .spyOn(SetupFileWatcher, "SetupFileWatcher")
+        .mockImplementationOnce(() => Promise.resolve());
 
       await LocationIsRemoteCallback(event, true);
 
@@ -65,6 +80,23 @@ describe("ipc bridge", () => {
 
     it("set to false", async () => {
       const event = { reply: jest.fn() } as any;
+
+      jest
+        .spyOn(appConfig, "set")
+        .mockReset()
+        .mockImplementationOnce(() => {
+          return Promise.resolve();
+        });
+
+      jest
+        .spyOn(createMainWindow, "default")
+        .mockImplementationOnce(() =>
+          Promise.resolve({ once: jest.fn() } as any)
+        );
+
+      jest
+        .spyOn(SetupFileWatcher, "SetupFileWatcher")
+        .mockImplementationOnce(() => Promise.resolve());
 
       jest.spyOn(appConfig, "get").mockImplementationOnce(() => {
         return Promise.resolve(false);
@@ -160,11 +192,16 @@ describe("ipc bridge", () => {
         .mockReset()
         .mockImplementationOnce(() => {
           return Promise.resolve();
+        })
+        .mockImplementationOnce(() => {
+          return Promise.resolve();
         });
 
       jest
         .spyOn(createMainWindow, "default")
-        .mockImplementationOnce(() => null);
+        .mockImplementationOnce(() =>
+          Promise.resolve({ once: jest.fn() } as any)
+        );
 
       jest
         .spyOn(SetupFileWatcher, "SetupFileWatcher")