+++ starskyapp/src/app/ipc-bridge/ipc-bridge.ts
@@ -16,6 +16,7 @@ import {
   LocationIsRemoteSettingsKey,
   LocationUrlSettingsKey
 } from "../config/location-settings.const";
+import RememberUrl from "../config/remember-url-settings.const";
 import { UpdatePolicyIpcKey } from "../config/update-policy-ipc-key.const";
 import { UpdatePolicySettings } from "../config/update-policy-settings.const";
 import UrlQuery from "../config/url-query";
@@ -25,6 +26,7 @@ import { SetupFileWatcher } from "../file-watcher/setup-file-watcher";
 import createMainWindow from "../main-window/create-main-window";
 import { mainWindows } from "../main-window/main-windows.const";
 import { GetNetRequest } from "../net-request/get-net-request";
+import { settingsWindows } from "../settings-window/settings-windows.const";
 
 function ipcBridge() {
   // When adding a new key also update preload-main.ts
@@ -77,7 +79,9 @@ export async function LocationIsRemoteCallback(
   args: boolean
 ) {
   if (args !== undefined && args !== null) {
-    await appConfig.set(LocationIsRemoteSettingsKey, args);
+    await SetupFileWatcher();
+    await closeAndCreateNewWindow();
+    await appConfig.set(LocationIsRemoteSettingsKey, args.toString());
   }
 
   const currentSettings = await appConfig.get(LocationIsRemoteSettingsKey);
@@ -134,12 +138,7 @@ export async function LocationUrlCallback(
         await appConfig.set(LocationUrlSettingsKey, locationUrl);
         // so you can save change the location
         await SetupFileWatcher();
-
-        // to avoid that the session is opened
-        mainWindows.forEach((window) => {
-          window.close();
-        });
-        createMainWindow("");
+        await closeAndCreateNewWindow();
       }
 
       console.log("locationOk >");
@@ -163,6 +162,22 @@ export async function LocationUrlCallback(
     isLocal: false,
     location: args
   } as IlocationUrlSettings);
+}
+
+/**
+ * to avoid that the session is opened
+ */
+async function closeAndCreateNewWindow() {
+  await appConfig.set(RememberUrl, {});
+  mainWindows.forEach((window) => {
+    window.close();
+  });
+  const newWindow = await createMainWindow("");
+  newWindow.once("ready-to-show", () => {
+    settingsWindows.forEach((window) => {
+      window.show();
+    });
+  });
 }
 
 export async function UpdatePolicyCallback(