+++ history.md
@@ -43,13 +43,19 @@ node starsky-tools/build-tools/app-version-update.js
 - change default option in thumbnailer cli to scan directories to enabled
 - remove newtonsoft.json references
 
-# version 0.4.4 _(Unreleased)_ - 2021-01-??
+# version 0.4.5 _(Unreleased)_ - 2021-01-??
+- nothing yet
+
+# version 0.4.4 - 2021-01-09
 - [x]   (Security) _App_ npm audit fix node_modules/ini 1.3.8
 - [x]   (Fixed) _Front-end_ When updating tags in sidebar and refresh afterwards its now not the old value anymore
 - [x]   (Fixed) _App_ Starting from another user in Mac OS should work now
 - [x]   (Fixed) _Front-end_ Label copy (press c and v) does not save with titles - issue #248
 - [x]   (Fixed) _Front-end_ update websocket data for other items outside view when receiving data bug - issue #265
-
+- [x]   (Fixed) _App_ when changing from local to remote it should open new window - issue #271
+- [x]   (Fixed) _App_ when changing from local to remote it should add watcher be updated - issue #271
+- [x]   (Fixed) _Front-end_ hide 'sign in instead' button on register page when there are no users yet
+  
 # version 0.4.3 - 2020-12-24
 - [x]   (Fixed) _Back-end_ the latest version isn't checked right, it takes the oldest version to compare with
 - [x]   (Changed) _App_ Rewrite of desktop application